Remove rpc api implementation
