--- node/src/main/scala/node/api/grpc/methods/All.scala
@@ -1,13 +1,0 @@
-package node.api.grpc.methods
-
-import io.grpc.ServerServiceDefinition
-import node.api.grpc.methods.Balances.{balancesMethodDescriptor, balancesServerCallHandler}
-
-object All {
-
-  /// Service (API) description, mapping of methods specification with handlers
-  def apply(serviceName: String, getBalance: String => Long): ServerServiceDefinition = ServerServiceDefinition
-    .builder(serviceName)
-    .addMethod(balancesMethodDescriptor, balancesServerCallHandler(getBalance))
-    .build()
-}
